added new solver parameter 'export_vtk' that controls whether VTK files are generated during workflow moved 'make_output_vtk_files' back to specfem base class so that is accessble by both 3d and 3d_globe added vtk file making to finalize procedures of base specfem
